[refactor] 바텀시트 오버레이 관련 해매는 중...
홈 화면 Surface -> Scaffold
