+++ feature/home/src/main/kotlin/com/nexters/bandalart/feature/home/BandalartBottomSheet.kt
@@ -202,7 +202,9 @@ fun BandalartBottomSheetContent(
             BottomSheetTopBar(
                 cellType = cellType,
                 isBlankCell = isBlankCell,
-                eventSink = eventSink,
+                onCloseClick = {
+                    eventSink(Event.ToggleCellBottomSheet(false))
+                }
             )
             Box {
                 Column(
