+++ feature/home/src/main/kotlin/com/nexters/bandalart/feature/home/HomeScreen.kt
@@ -4,21 +4,50 @@ import android.app.Activity
 import android.content.Context
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.animateContentSize
+import androidx.compose.animation.core.LinearOutSlowInEasing
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarDuration.Indefinite
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SnackbarResult
-import androidx.compose.material3.Surface
+import androidx.compose.material3.Switch
+import androidx.compose.material3.SwitchDefaults
+import androidx.compose.material3.Text
+import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -27,12 +56,21 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithContent
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Color.Companion.Transparent
 import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.layer.drawLayer
 import androidx.compose.ui.graphics.rememberGraphicsLayer
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.google.android.play.core.appupdate.AppUpdateManagerFactory
 import com.google.android.play.core.appupdate.AppUpdateOptions
 import com.google.android.play.core.install.InstallStateUpdatedListener
@@ -40,17 +78,30 @@ import com.google.android.play.core.install.model.AppUpdateType
 import com.google.android.play.core.install.model.InstallStatus
 import com.google.android.play.core.install.model.UpdateAvailability
 import com.nexters.bandalart.core.common.extension.await
+import com.nexters.bandalart.core.common.extension.clearFocusOnKeyboardDismiss
+import com.nexters.bandalart.core.common.extension.getCurrentLocale
+import com.nexters.bandalart.core.common.extension.noRippleClickable
+import com.nexters.bandalart.core.common.extension.toLocalDateTime
+import com.nexters.bandalart.core.common.extension.toStringLocalDateTime
 import com.nexters.bandalart.core.common.utils.UiText
 import com.nexters.bandalart.core.common.utils.isValidImmediateAppUpdate
 import com.nexters.bandalart.core.designsystem.theme.BandalartTheme
+import com.nexters.bandalart.core.designsystem.theme.BottomSheetContent
 import com.nexters.bandalart.core.designsystem.theme.Gray100
+import com.nexters.bandalart.core.designsystem.theme.Gray300
+import com.nexters.bandalart.core.designsystem.theme.Gray400
 import com.nexters.bandalart.core.designsystem.theme.Gray50
+import com.nexters.bandalart.core.designsystem.theme.Gray700
+import com.nexters.bandalart.core.designsystem.theme.White
 import com.nexters.bandalart.core.domain.entity.BandalartCellEntity
 import com.nexters.bandalart.core.domain.entity.UpdateBandalartEmojiEntity
 import com.nexters.bandalart.core.ui.DevicePreview
+import com.nexters.bandalart.core.ui.NavigationBarHeightDp
 import com.nexters.bandalart.core.ui.R
-import com.nexters.bandalart.feature.home.ui.bandalart.BandalartDeleteAlertDialog
+import com.nexters.bandalart.core.ui.ThemeColor
+import com.nexters.bandalart.core.ui.getNavigationBarPadding
 import com.nexters.bandalart.feature.home.HomeScreen.Event
+import com.nexters.bandalart.feature.home.HomeScreen.State
 import com.nexters.bandalart.feature.home.model.BandalartUiModel
 import com.nexters.bandalart.feature.home.model.CellType
 import com.nexters.bandalart.feature.home.model.dummy.dummyBandalartChartData
@@ -60,17 +111,27 @@ import com.nexters.bandalart.feature.home.ui.HomeHeader
 import com.nexters.bandalart.feature.home.ui.HomeShareButton
 import com.nexters.bandalart.feature.home.ui.HomeTopBar
 import com.nexters.bandalart.feature.home.ui.bandalart.BandalartChart
+import com.nexters.bandalart.feature.home.ui.bandalart.BandalartColorPicker
+import com.nexters.bandalart.feature.home.ui.bandalart.BandalartDatePicker
+import com.nexters.bandalart.feature.home.ui.bandalart.BandalartDeleteAlertDialog
 import com.nexters.bandalart.feature.home.ui.bandalart.BandalartEmojiBottomSheet
+import com.nexters.bandalart.feature.home.ui.bandalart.BandalartEmojiPicker
 import com.nexters.bandalart.feature.home.ui.bandalart.BandalartListBottomSheet
 import com.nexters.bandalart.feature.home.ui.bandalart.BandalartSkeleton
+import com.nexters.bandalart.feature.home.ui.bandalart.BottomSheetCompleteButton
+import com.nexters.bandalart.feature.home.ui.bandalart.BottomSheetContentPlaceholder
+import com.nexters.bandalart.feature.home.ui.bandalart.BottomSheetContentText
+import com.nexters.bandalart.feature.home.ui.bandalart.BottomSheetDeleteButton
+import com.nexters.bandalart.feature.home.ui.bandalart.BottomSheetSubTitleText
+import com.nexters.bandalart.feature.home.ui.bandalart.BottomSheetTopBar
 import com.nexters.bandalart.feature.home.viewmodel.ModalType
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.overlay.LocalOverlayHost
-import com.slack.circuit.overlay.Overlay
 import com.slack.circuit.overlay.OverlayEffect
 import com.slack.circuit.runtime.CircuitUiEvent
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.screen.Screen
+import com.slack.circuitx.overlays.BottomSheetOverlay
 import dagger.hilt.android.components.ActivityRetainedComponent
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
@@ -78,6 +139,7 @@ import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.launch
 import kotlinx.parcelize.Parcelize
 import timber.log.Timber
+import java.time.LocalDateTime
 
 // TODO Action 와 Event 를 분리하지 말고 바로 Event 를 호출하면 될듯 (이게 MVI 가 맞나?)
 @Parcelize
@@ -99,6 +161,8 @@ data object HomeScreen : Screen {
         val clickedCellType: CellType = CellType.MAIN,
         val clickedCellData: BandalartCellEntity? = null,
         val isUpdateAlreadyRejected: Boolean = false,
+        val isEmojiPickerOpened: Boolean = false,
+        val isDatePickerOpened: Boolean = false,
         val eventSink: (Event) -> Unit,
     ) : CircuitUiState
 
@@ -120,6 +184,7 @@ data object HomeScreen : Screen {
         data object OnListClick : Event
         data object OnSaveClick : Event
         data object OnDeleteClick : Event
+        data object OnEmojiClick: Event
         data class OnEmojiSelected(
             val bandalartId: Long,
             val cellId: Long,
@@ -144,6 +209,7 @@ data object HomeScreen : Screen {
 
         data object OnCloseButtonClick : Event
         data object OnAppTitleClick : Event
+        data object OnDatePickerClick : Event
     }
 }
 
@@ -295,10 +361,11 @@ data object HomeScreen : Screen {
 //}
 
 // TODO Scaffold 를 통해 NavHost 를 감싸지 않았기 때문에, 모든 화면에 Scaffold 를 래핑해줘야 할 듯 하다. (innerPadding 을 적용하기 위해)
+@OptIn(ExperimentalMaterial3Api::class)
 @CircuitInject(HomeScreen::class, ActivityRetainedComponent::class)
 @Composable
 internal fun Home(
-    state: HomeScreen.State,
+    state: State,
 //    onHomeUiAction: (HomeUiAction) -> Unit,
 //    shareBandalart: (ImageBitmap) -> Unit,
 //    captureBandalart: (ImageBitmap) -> Unit,
@@ -406,22 +473,15 @@ internal fun Home(
         }
     }
 
-    OverlayEffect {
-        if (state.isEmojiBottomSheetOpened) {
-            state.bandalartData?.let { bandalart ->
-                state.bandalartCellData?.let { cell ->
-                    Overlay(
-                        content = {
-                            BandalartEmojiBottomSheet(
-                                bandalartId = bandalart.id,
-                                cellId = cell.id,
-                                currentEmoji = bandalart.profileEmoji,
-                                eventSink = eventSink,
-                            )
-                        },
-                        overlayHost = overlayHost,
-                    )
-                }
+    if (state.isEmojiBottomSheetOpened) {
+        state.bandalartData?.let { bandalart ->
+            state.bandalartCellData?.let { cell ->
+                BandalartEmojiBottomSheet(
+                    bandalartId = bandalart.id,
+                    cellId = cell.id,
+                    currentEmoji = bandalart.profileEmoji,
+                    eventSink = eventSink,
+                )
             }
         }
     }
@@ -442,13 +502,383 @@ internal fun Home(
     if (state.isCellBottomSheetOpened) {
         state.bandalartData?.let { bandalart ->
             state.clickedCellData?.let { cell ->
-                BandalartBottomSheet(
-                    bandalartId = bandalart.id,
-                    cellType = state.clickedCellType,
-                    isBlankCell = cell.title.isNullOrEmpty(),
-                    cellData = cell,
-                    eventSink = eventSink,
-                )
+                OverlayEffect {
+                    val result = show(
+                        BottomSheetOverlay<BandalartBottomSheetModel, BandalartBottomSheetResult>(
+                            model = BandalartBottomSheetModel(
+                                bandalartId = bandalart.id,
+                                cellType = state.clickedCellType,
+                                isBlankCell = cell.title.isNullOrEmpty(),
+                                initialBandalartData = bandalart,
+                                bandalartData = bandalart,
+                                cellData = cell,
+                                initialCellData = cell,
+                            ),
+                            onDismiss = { BandalartBottomSheetResult.Dismiss },
+                            // onDismiss = { Navigator.NoOp },
+                            skipPartiallyExpandedState = true,
+                        ) { model, navigator ->
+                            val context = LocalContext.current
+                            val scope = rememberCoroutineScope()
+                            val bottomSheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
+                            val focusManager = LocalFocusManager.current
+                            val scrollState = rememberScrollState()
+                            val currentLocale = context.getCurrentLocale()
+
+                            Column(
+                                modifier = Modifier
+                                    .background(White)
+                                    .navigationBarsPadding()
+                                    .noRippleClickable { focusManager.clearFocus() },
+                            ) {
+                                Spacer(modifier = Modifier.height(20.dp))
+                                BottomSheetTopBar(
+                                    cellType = model.cellType,
+                                    isBlankCell = model.isBlankCell,
+                                    onCloseClick = {
+                                        eventSink(Event.ToggleCellBottomSheet(false))
+                                    }
+                                )
+                                Box {
+                                    Column(
+                                        modifier = Modifier
+                                            .fillMaxWidth()
+                                            .wrapContentHeight()
+                                            .verticalScroll(scrollState)
+                                            .padding(start = 20.dp, top = 40.dp, end = 20.dp),
+                                    ) {
+                                        BottomSheetSubTitleText(text = stringResource(R.string.bottomsheet_title))
+                                        Spacer(modifier = Modifier.height(11.dp))
+                                        Row(modifier = Modifier.fillMaxWidth()) {
+                                            if (model.cellType == CellType.MAIN) {
+                                                Box(
+                                                    modifier = Modifier
+                                                        .align(Alignment.CenterVertically)
+                                                        .padding(end = 16.dp),
+                                                ) {
+                                                    Card(
+                                                        shape = RoundedCornerShape(16.dp),
+                                                        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp),
+                                                    ) {
+                                                        Box(
+                                                            modifier = Modifier
+                                                                .width(52.dp)
+                                                                .aspectRatio(1f)
+                                                                .background(Gray100)
+                                                                .clickable {
+                                                                    eventSink(Event.OnEmojiClick)
+                                                                },
+                                                            contentAlignment = Alignment.Center,
+                                                        ) {
+                                                            if (model.bandalartData.profileEmoji.isNullOrEmpty()) {
+                                                                Icon(
+                                                                    imageVector = ImageVector.vectorResource(com.nexters.bandalart.core.designsystem.R.drawable.ic_empty_emoji),
+                                                                    contentDescription = stringResource(R.string.empty_emoji_description),
+                                                                    tint = Color.Unspecified,
+                                                                )
+                                                            } else {
+                                                                Text(
+                                                                    text = model.bandalartData.profileEmoji,
+                                                                    fontSize = 22.sp,
+                                                                )
+                                                            }
+                                                        }
+                                                    }
+                                                    Icon(
+                                                        imageVector = ImageVector.vectorResource(com.nexters.bandalart.core.designsystem.R.drawable.ic_edit),
+                                                        contentDescription = stringResource(R.string.edit_description),
+                                                        tint = Color.Unspecified,
+                                                        modifier = Modifier
+                                                            .align(Alignment.BottomEnd)
+                                                            .offset(x = 4.dp, y = 4.dp),
+                                                    )
+                                                }
+                                            }
+                                            Column(modifier = Modifier.padding(top = 10.dp)) {
+                                                BasicTextField(
+                                                    value = model.cellData.title ?: "",
+                                                    onValueChange = { title ->
+                                                        navigator.finish(BandalartBottomSheetResult.OnTitleUpdate(title, currentLocale))
+                                                    },
+                                                    modifier = Modifier
+                                                        .fillMaxWidth()
+                                                        .height(24.dp)
+                                                        .clearFocusOnKeyboardDismiss(),
+                                                    keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Done),
+                                                    keyboardActions = KeyboardActions(onDone = { focusManager.clearFocus() }),
+                                                    maxLines = 1,
+                                                    textStyle = BottomSheetContent,
+                                                    decorationBox = { innerTextField ->
+                                                        if (model.cellData.title.isNullOrEmpty()) {
+                                                            BottomSheetContentPlaceholder(text = stringResource(R.string.bottomsheet_title_placeholder))
+                                                        }
+                                                        innerTextField()
+                                                    },
+                                                )
+                                                Spacer(modifier = Modifier.height(10.dp))
+                                                HorizontalDivider(
+                                                    modifier = Modifier.fillMaxWidth(),
+                                                    thickness = 1.dp,
+                                                    color = Gray300,
+                                                )
+                                            }
+                                        }
+                                        AnimatedVisibility(visible = state.isEmojiPickerOpened) {
+                                            Column {
+                                                BandalartEmojiPicker(
+                                                    modifier = Modifier
+                                                        .wrapContentSize()
+                                                        .padding(top = 4.dp)
+                                                        .animateContentSize(
+                                                            animationSpec = tween(
+                                                                durationMillis = 300,
+                                                                easing = LinearOutSlowInEasing,
+                                                            ),
+                                                        ),
+                                                    currentEmoji = model.bandalartData.profileEmoji,
+                                                    isBottomSheet = false,
+                                                    onEmojiSelect = { selectedEmoji ->
+                                                        navigator.finish(BandalartBottomSheetResult.OnEmojiSelect(selectedEmoji))
+                                                    },
+                                                )
+                                            }
+                                        }
+                                        if (model.cellType == CellType.MAIN) {
+                                            Spacer(modifier = Modifier.height(22.dp))
+                                            BottomSheetSubTitleText(text = stringResource(R.string.bottomsheet_color))
+                                            BandalartColorPicker(
+                                                initColor = ThemeColor(
+                                                    mainColor = model.bandalartData.mainColor,
+                                                    subColor = model.bandalartData.subColor,
+                                                ),
+                                                onColorSelect = {
+                                                    navigator.finish(BandalartBottomSheetResult.OnColorSelect(it.mainColor, it.subColor))
+                                                },
+                                            )
+                                            Spacer(modifier = Modifier.height(3.dp))
+                                        }
+                                        Spacer(modifier = Modifier.height(25.dp))
+                                        BottomSheetSubTitleText(text = stringResource(R.string.bottomsheet_duedate))
+                                        Spacer(modifier = Modifier.height(12.dp))
+                                        Column {
+                                            Box(
+                                                modifier = Modifier
+                                                    .fillMaxWidth()
+                                                    .height(24.dp)
+                                                    .clickable {
+                                                        // eventSink(Event.OnDatePickerClick)
+                                                        navigator.finish(BandalartBottomSheetResult.OnDatePickerClick)
+                                                    },
+                                                contentAlignment = Alignment.CenterStart,
+                                            ) {
+                                                if (model.cellData.dueDate.isNullOrEmpty()) {
+                                                    BottomSheetContentPlaceholder(text = stringResource(R.string.bottomsheet_duedate_placeholder))
+                                                } else {
+                                                    BottomSheetContentText(text = model.cellData.dueDate!!.toStringLocalDateTime())
+                                                }
+                                                Icon(
+                                                    imageVector = Icons.AutoMirrored.Filled.KeyboardArrowRight,
+                                                    contentDescription = stringResource(R.string.arrow_forward_description),
+                                                    modifier = Modifier
+                                                        .align(Alignment.CenterEnd)
+                                                        .size(24.dp),
+                                                    tint = Gray400,
+                                                )
+                                            }
+                                            Spacer(modifier = Modifier.height(10.dp))
+                                            HorizontalDivider(
+                                                modifier = Modifier.fillMaxWidth(),
+                                                thickness = 1.dp,
+                                                color = Gray300,
+                                            )
+                                        }
+                                        AnimatedVisibility(visible = state.isDatePickerOpened) {
+                                            BandalartDatePicker(
+                                                onDueDateSelect = { dueDateResult ->
+                                                    navigator.finish(BandalartBottomSheetResult.OnDueDateSelect(dueDateResult.toString()))
+                                                },
+                                                currentDueDate = model.cellData.dueDate?.toLocalDateTime() ?: LocalDateTime.now(),
+                                            )
+                                        }
+                                        Spacer(modifier = Modifier.height(28.dp))
+                                        BottomSheetSubTitleText(text = stringResource(R.string.bottomsheet_description))
+                                        Spacer(modifier = Modifier.height(12.dp))
+                                        Box {
+                                            Column {
+                                                BasicTextField(
+                                                    value = model.cellData.description ?: "",
+                                                    onValueChange = { description ->
+                                                        navigator.finish(BandalartBottomSheetResult.OnDescriptionUpdate(description))
+                                                    },
+                                                    modifier = Modifier
+                                                        .fillMaxWidth()
+                                                        .height(24.dp)
+                                                        .clearFocusOnKeyboardDismiss(),
+                                                    keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Done),
+                                                    keyboardActions = KeyboardActions(onDone = { focusManager.clearFocus() }),
+                                                    maxLines = 1,
+                                                    textStyle = BottomSheetContent,
+                                                    decorationBox = { innerTextField ->
+                                                        if (model.cellData.description.isNullOrEmpty()) {
+                                                            BottomSheetContentPlaceholder(text = stringResource(R.string.bottomsheet_description_placeholder))
+                                                        }
+                                                        innerTextField()
+                                                    },
+                                                )
+                                                Spacer(modifier = Modifier.height(10.dp))
+                                                HorizontalDivider(
+                                                    modifier = Modifier.fillMaxWidth(),
+                                                    thickness = 1.dp,
+                                                    color = Gray300,
+                                                )
+                                            }
+                                        }
+                                        if (model.cellType == CellType.TASK) {
+                                            Spacer(modifier = Modifier.height(28.dp))
+                                            BottomSheetSubTitleText(text = stringResource(R.string.bottomsheet_is_completed))
+                                            Spacer(modifier = Modifier.height(12.dp))
+                                            Box(modifier = Modifier.fillMaxWidth()) {
+                                                BottomSheetContentText(
+                                                    modifier = Modifier.align(Alignment.CenterStart),
+                                                    text = if (model.cellData.isCompleted) stringResource(R.string.bottomsheet_completed)
+                                                    else stringResource(R.string.bottomsheet_in_completed),
+                                                )
+                                                Switch(
+                                                    checked = model.cellData.isCompleted,
+                                                    onCheckedChange = { isCompleted ->
+                                                        navigator.finish(BandalartBottomSheetResult.OnCompletionUpdate(isCompleted))
+                                                    },
+                                                    colors = SwitchDefaults.colors(
+                                                        uncheckedThumbColor = White,
+                                                        uncheckedTrackColor = Gray300,
+                                                        uncheckedBorderColor = Gray300,
+                                                        checkedThumbColor = White,
+                                                        checkedTrackColor = Gray700,
+                                                        checkedBorderColor = Gray700,
+                                                    ),
+                                                    modifier = Modifier
+                                                        .align(Alignment.CenterEnd)
+                                                        .width(52.dp)
+                                                        .height(28.dp),
+                                                )
+                                            }
+                                            Spacer(modifier = Modifier.height(4.dp))
+                                        }
+                                        Spacer(modifier = Modifier.height(28.dp))
+                                        Row(
+                                            modifier = Modifier
+                                                .fillMaxWidth()
+                                                .align(Alignment.CenterHorizontally)
+                                                .padding(horizontal = 8.dp)
+                                                .imePadding(),
+                                        ) {
+                                            if (!model.isBlankCell) {
+                                                BottomSheetDeleteButton(
+                                                    onClick = {
+                                                        navigator.finish(BandalartBottomSheetResult.OnDeleteButtonClick)
+                                                    },
+                                                    modifier = Modifier.weight(1f),
+                                                )
+                                                Spacer(modifier = Modifier.width(9.dp))
+                                            }
+                                            BottomSheetCompleteButton(
+                                                isEnabled = (model.cellData.title?.trim()
+                                                    ?.isNotEmpty() == true) && (model.cellData != model.initialCellData || model.bandalartData != model.initialBandalartData),
+                                                onClick = {
+                                                    navigator.finish(BandalartBottomSheetResult.OnCompleteButtonClick(model.bandalartId, model.cellData.id, model.cellType))
+                                                },
+                                                modifier = Modifier.weight(1f),
+                                            )
+                                        }
+                                        Spacer(modifier = Modifier.height(NavigationBarHeightDp + getNavigationBarPadding()))
+                                    }
+                                    if (scrollState.value > 0) {
+                                        Column(
+                                            modifier = Modifier
+                                                .background(
+                                                    brush = Brush.verticalGradient(
+                                                        colors = listOf(White, Transparent),
+                                                    ),
+                                                    shape = RectangleShape,
+                                                )
+                                                .height(77.dp)
+                                                .fillMaxWidth(),
+                                        ) {}
+                                    }
+                                    if (scrollState.value < scrollState.maxValue) {
+                                        Column(
+                                            modifier = Modifier
+                                                .background(
+                                                    brush = Brush.verticalGradient(
+                                                        colors = listOf(Transparent, White),
+                                                    ),
+                                                    shape = RectangleShape,
+                                                )
+                                                .height(77.dp)
+                                                .fillMaxWidth()
+                                                .align(Alignment.BottomCenter),
+                                        ) {}
+                                    }
+                                }
+                            }
+                        },
+                    )
+
+                    // 결과 처리
+                    when (result) {
+                        is BandalartBottomSheetResult.Dismiss -> {
+                            state.eventSink(Event.ToggleCellBottomSheet(false))
+                        }
+
+                        is BandalartBottomSheetResult.OnCompleteButtonClick -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnDeleteButtonClick -> {
+
+                        }
+
+                        is BandalartBottomSheetResult. OnDeleteDialogConfirmClick -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnDeleteDialogCancelClick -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnDatePickerClick -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnEmojiPickerClick -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnEmojiSelect -> {
+                            eventSink(Event.OnEmojiClick)
+                        }
+
+                        is BandalartBottomSheetResult.OnColorSelect -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnDueDateSelect -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnTitleUpdate -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnDescriptionUpdate -> {
+
+                        }
+
+                        is BandalartBottomSheetResult.OnCompletionUpdate -> {
+
+                        }
+                    }
+                }
             }
         }
     }
@@ -469,11 +899,14 @@ internal fun Home(
         }
     }
 
-    Surface(
+    Scaffold(
         modifier = modifier.fillMaxSize(),
-        color = Gray50,
-    ) {
-        Box(modifier = Modifier.fillMaxSize()) {
+        containerColor = Gray50,
+    ) { innerPadding ->
+        Box(modifier = Modifier
+            .fillMaxSize()
+            .padding(innerPadding)
+        ) {
             Column(
                 modifier = Modifier
                     .fillMaxSize()
@@ -565,11 +998,11 @@ private fun updateBandalartListTitles(
 private fun HomeScreenSingleBandalartPreview() {
     BandalartTheme {
         Home(
-            state = HomeScreen.State(
+            state = State(
                 bandalartList = listOf(dummyBandalartList[0]).toImmutableList(),
                 bandalartData = dummyBandalartData,
                 bandalartCellData = dummyBandalartChartData,
-                eventSink = {}
+                eventSink = {},
             ),
         )
     }
@@ -580,11 +1013,11 @@ private fun HomeScreenSingleBandalartPreview() {
 private fun HomeScreenMultipleBandalartPreview() {
     BandalartTheme {
         Home(
-            state = HomeScreen.State(
+            state = State(
                 bandalartList = dummyBandalartList.toImmutableList(),
                 bandalartData = dummyBandalartData,
                 bandalartCellData = dummyBandalartChartData,
-                eventSink = {}
+                eventSink = {},
             ),
         )
     }
